Generate C++ code with curated parameters

+++ pencode_core.m
@@ -6,8 +6,8 @@ for i=1:n
     nB = 2^(i-1);
     for j=1:nB
         base = (j-1)*B;
-        for l=1:B/2
-            x(base+l) = mod( x(base+l)+x(base+B/2+l), 2 );
+        for l=1:B/int32(2)
+            x(base+l) = mod( x(base+l)+x(base+B/int32(2)+l), int32(2) );
         end
     end
 end
